added different config to allow for tests later
