--- public/manifest.json
@@ -1,17 +1,0 @@
-{
-  "name": "dark-mode",
-  "description": "A Chrome extension for dark mode on specific websites. Inspired by the brightness of email calendars.",
-  "version": "1.0.0",
-  "manifest_version": 3,
-	"permissions": [
-		"activeTab",
-		"scripting"
-	],
-  "action": {
-    "default_title": "Dark Mode",
-    "default_popup": "background.tsx"
-  },
-	"background": {
-		"service_worker": "background.js"
-	}
-}
