+++ src/App.css
@@ -1,0 +1,11 @@
+.App {
+  color: white;
+  text-align: center;
+}
+
+.container {
+  width: 15rem;
+  height: 15rem;
+  background-color: green;
+
+}
