+++ tsconfig.json
@@ -1,14 +1,26 @@
 {
-	"compilerOptions": {
-		"strict": true,
-		"module": "commonjs",
-		"jsx": "react",
-		"target": "es6",
-		"esModuleInterop": true,
-		"sourceMap": true,
-		"rootDir": "src",
-		"outDir": "dist/js",
-		"noEmitOnError": true,
-		"typeRoots": [ "node_modules/@types" ]
-	}
+  "compilerOptions": {
+    "target": "es5",
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
+    "allowJs": true,
+    "skipLibCheck": true,
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "strict": true,
+    "forceConsistentCasingInFileNames": true,
+    "noFallthroughCasesInSwitch": true,
+    "module": "esnext",
+    "moduleResolution": "node",
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "noEmit": true,
+    "jsx": "react-jsx"
+  },
+  "include": [
+    "src"
+  ]
 }
