--- webpack/webpack.config.js
@@ -1,29 +1,0 @@
-const path = require('path');
-const CopyPlugin = require('copy-webpack-plugin');
-module.exports = {
-   mode: "production",
-   entry: {
-      background: path.resolve(__dirname, "..", "src", "background.tsx"),
-   },
-   output: {
-      path: path.join(__dirname, "../dist"),
-      filename: "[name].js",
-   },
-   resolve: {
-      extensions: [".ts", ".js", ".jsx"],
-   },
-   module: {
-      rules: [
-         {
-            test: /\.tsx?$/,
-            loader: "ts-loader",
-            exclude: /node_modules/,
-         },
-      ],
-   },
-   plugins: [
-      new CopyPlugin({
-         patterns: [{from: ".", to: ".", context: "public"}]
-      }),
-   ],
-};
